Feature/saved list be develop

+++ core-customize/tccore/src/com/taloscommerce/core/customproductlist/impl/DefaultCustomProductListService.java
@@ -64,6 +64,9 @@ public class DefaultCustomProductListService implements CustomProductListService
                 productListModel.setCustomer(customer);
                 getModelService().save(productListModel);
             }
+            else {
+                throw new IllegalArgumentException("List with Name " +productListModel.getName() +" already exists");
+            }
         }
         return productListModel;
     }
@@ -93,6 +96,9 @@ public class DefaultCustomProductListService implements CustomProductListService
     @Override
     public void deleteCustomProductList(final String customProductListId) {
         final CustomProductListModel model = getCustomProductListById(customProductListId);
+        if (Objects.isNull(model)){
+            throw new UnknownIdentifierException("CustomProductList with Id "+customProductListId+" can not be removed, it did not exist");
+        }
         getModelService().remove(model);
     }
 
@@ -107,6 +113,9 @@ public class DefaultCustomProductListService implements CustomProductListService
                 productModels.remove(productModel);
                 productListModel.setProduct(productModels);
                 getModelService().save(productListModel);
+            }
+            else {
+                throw new IllegalArgumentException("List did not contain product "+productCode);
             }
         }
     }
